# src/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TicketCRM API",
        "description": "API documentation for TicketCRM application",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8103",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/tickets": {
            "get": {
                "tags": [
                    "Authenticated Tickets"
                ],
                "summary": "Get all tickets",
                "description": "Returns a list of all tickets for authenticated user",
                "operationId": "ddc3a5e66a83482e874c84e71ebc4505",
                "responses": {
                    "200": {
                        "description": "List of tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ticket"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Authenticated Tickets"
                ],
                "summary": "Create a new ticket",
                "description": "Creates a new ticket with customer information (authenticated users)",
                "operationId": "cfc7c1ca165d45fb488443df73e66587",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "theme",
                                    "text",
                                    "email"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "description": "ID of existing customer (optional if providing email/phone)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "description": "Customer email (required if no customer_id provided)",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "description": "Customer phone (optional)",
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "name": {
                                        "description": "Customer name (optional)",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "theme": {
                                        "description": "Ticket theme/subject",
                                        "type": "string",
                                        "example": "Support Request"
                                    },
                                    "text": {
                                        "description": "Ticket content",
                                        "type": "string",
                                        "example": "Detailed description of the issue"
                                    },
                                    "status": {
                                        "description": "Ticket status",
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "in_progress",
                                            "done"
                                        ],
                                        "example": "new"
                                    },
                                    "date_answer": {
                                        "description": "Date of answer",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-25T10:00Z"
                                    },
                                    "attachments[]": {
                                        "description": "Attachments to attach to the ticket",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket created successfully"
                                        },
                                        "ticket": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/feedback": {
            "post": {
                "tags": [
                    "Public Tickets"
                ],
                "summary": "Create a ticket from feedback form",
                "description": "Creates a new ticket from the feedback widget form (public endpoint)",
                "operationId": "4675803eb9b8c476b49e4793148ed9d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "theme",
                                    "text",
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Customer email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "description": "Customer phone (optional)",
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "name": {
                                        "description": "Customer name (optional)",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "theme": {
                                        "description": "Ticket theme/subject",
                                        "type": "string",
                                        "example": "Feedback"
                                    },
                                    "text": {
                                        "description": "Ticket content",
                                        "type": "string",
                                        "example": "Detailed feedback message"
                                    },
                                    "attachments[]": {
                                        "description": "Attachments to attach to the ticket",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully from feedback",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket created successfully"
                                        },
                                        "ticket": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/tickets/{ticket}": {
            "get": {
                "tags": [
                    "Authenticated Tickets"
                ],
                "summary": "Get a specific ticket",
                "description": "Returns details of a specific ticket",
                "operationId": "6d7ec6f4fe57a18b4f2f799a9d74e3fb",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tickets/{ticket}/status": {
            "patch": {
                "tags": [
                    "Authenticated Tickets"
                ],
                "summary": "Update ticket status",
                "description": "Updates the status of a specific ticket",
                "operationId": "2126147b70e1171e7a56f3f063ca9e9f",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "New ticket status",
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "in_progress",
                                            "done"
                                        ],
                                        "example": "in_progress"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket status updated successfully"
                                        },
                                        "ticket": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Ticket not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Ticket": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer": {
                        "$ref": "#/components/schemas/Customer"
                    },
                    "theme": {
                        "type": "string",
                        "example": "Support Request"
                    },
                    "text": {
                        "type": "string",
                        "example": "Detailed description of the issue"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "in_progress",
                            "done"
                        ],
                        "example": "new"
                    },
                    "date_answer": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-25T10:00Z"
                    },
                    "files": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/File"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}